requirements.txt
albumentations
torch
pandas
numpy
pillow
scipy
matplotlib
torchvision
wandb
sklearn
streamlit
tqdm